requirements.txt
fastapi
uvicorn
sqlmodel
google-genai
python-dotenv
requests